fix: remove unnecessary libraries and rename local storage utilitie
