+++ package.json
@@ -19,7 +19,6 @@
     "@fortawesome/free-regular-svg-icons": "^6.4.2",
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "add": "^2.0.6",
     "babel-loader": "^9.1.3",
     "css-loader": "^6.8.1",
     "file-loader": "^6.2.0",
@@ -37,8 +36,7 @@
     "webpack-bundle-tracker": "^2.0.1",
     "webpack-cli": "^5.1.4",
     "webpack-manifest-plugin": "^5.0.0",
-    "webpack-merge": "^5.9.0",
-    "yarn": "^1.22.19"
+    "webpack-merge": "^5.9.0"
   },
   "devDependencies": {
     "babel-plugin-module-resolver": "^5.0.0",
